Chore: Added Console to remove mixed functionality from LogLayer
